Refactor scene view tests
